#5 moved Album-Query to it's own class
