+++ code/src/database/MySQLArtistReader.php
@@ -41,23 +41,4 @@ class MySQLArtistReader
 
         return $result;
     }
-
-    public function getAlbum(int|null $albumId): array
-    {
-        if ($albumId === null){
-            return [];
-        }
-
-        $sql = $this->mySQLConnection->prepare('
-            SELECT *
-              FROM Album 
-            WHERE Album.id = :album_id
-        ');
-
-        $sql->bindValue(":album_id", $albumId);
-        $sql->execute();
-        $result = $sql->fetch(PDO::FETCH_ASSOC);
-
-        return $result;
-    }
 }
